(untangle) added simulation.tick(50) call, after restoring the stashed properties.
this produces less time the user waits and watches as the simulation cools, but leaves a brief duration when the cooling is finishing, for the user to see.

in this commit also extensively added comments, elucidating some less obvious parts of the code.

in this commit also renamed a paramater of the notebook's export function, for clarity, to 'observerFactory'. also removed observerFactory arguments to main.variable() calls when not needed.